+++ mempool/mempool.go
@@ -30,7 +30,12 @@ import (
 
 var _ sdkmempool.ExtMempool = &ExperimentalEVMMempool{}
 
-const SubscriberName = "evm"
+const (
+	// SubscriberName is the name of the event bus subscriber for the EVM mempool
+	SubscriberName = "evm"
+	// fallbackBlockGasLimit is the default block gas limit is 0 or missing in genesis file
+	fallbackBlockGasLimit = 100_000_000
+)
 
 type (
 	// ExperimentalEVMMempool is a unified mempool that manages both EVM and Cosmos SDK transactions.
@@ -53,6 +58,7 @@ type (
 		bondDenom     string
 		evmDenom      string
 		blockGasLimit uint64 // Block gas limit from consensus parameters
+		minTip        *uint256.Int
 
 		/** Verification **/
 		anteHandler sdk.AnteHandler
@@ -75,6 +81,7 @@ type EVMMempoolConfig struct {
 	AnteHandler      sdk.AnteHandler
 	BroadCastTxFn    func(txs []*ethtypes.Transaction) error
 	BlockGasLimit    uint64 // Block gas limit from consensus parameters
+	MinTip           *uint256.Int
 }
 
 // NewExperimentalEVMMempool creates a new unified mempool for EVM and Cosmos transactions.
@@ -83,10 +90,8 @@ type EVMMempoolConfig struct {
 // of pools and verification functions, with sensible defaults created if not provided.
 func NewExperimentalEVMMempool(getCtxCallback func(height int64, prove bool) (sdk.Context, error), logger log.Logger, vmKeeper VMKeeperI, feeMarketKeeper FeeMarketKeeperI, txConfig client.TxConfig, clientCtx client.Context, config *EVMMempoolConfig) *ExperimentalEVMMempool {
 	var (
-		txPool      *txpool.TxPool
-		cosmosPool  sdkmempool.ExtMempool
-		anteHandler sdk.AnteHandler
-		blockchain  *Blockchain
+		cosmosPool sdkmempool.ExtMempool
+		blockchain *Blockchain
 	)
 
 	bondDenom := evmtypes.GetEVMCoinDenom()
@@ -101,13 +106,12 @@ func NewExperimentalEVMMempool(getCtxCallback func(height int64, prove bool) (sd
 		panic("config must not be nil")
 	}
 
-	anteHandler = config.AnteHandler
-	blockchain = newBlockchain(getCtxCallback, logger, vmKeeper, feeMarketKeeper, config.BlockGasLimit)
-
 	if config.BlockGasLimit == 0 {
-		logger.Debug("block gas limit is 0, setting default", "default_limit", 100_000_000)
-		config.BlockGasLimit = 100_000_000
+		logger.Warn("block gas limit is 0, setting to fallback", "fallback_limit", fallbackBlockGasLimit)
+		config.BlockGasLimit = fallbackBlockGasLimit
 	}
+
+	blockchain = newBlockchain(getCtxCallback, logger, vmKeeper, feeMarketKeeper, config.BlockGasLimit)
 
 	// Create txPool from configuration
 	legacyConfig := legacypool.DefaultConfig
@@ -183,7 +187,7 @@ func NewExperimentalEVMMempool(getCtxCallback func(height int64, prove bool) (sd
 		bondDenom:     bondDenom,
 		evmDenom:      evmDenom,
 		blockGasLimit: config.BlockGasLimit,
-		anteHandler:   anteHandler,
+		minTip:        config.MinTip,
 		txGases:       make(map[string]uint64),
 	}
 
@@ -380,7 +384,7 @@ func (m *ExperimentalEVMMempool) shouldRemoveFromEVMPool(tx sdk.Tx) bool {
 
 	// If it was a successful transaction or a sequence error, we let the mempool handle the cleaning.
 	// If it was any other Cosmos or antehandler related issue, then we remove it.
-	ctx, err := m.blockchain.GetLatestCtx()
+	ctx, err := m.blockchain.GetLatestContext()
 	if err != nil {
 		m.logger.Debug("cannot get latest context for validation, keeping transaction", "error", err)
 		return false // Cannot validate, keep transaction
@@ -487,7 +491,7 @@ func (m *ExperimentalEVMMempool) getIterators(goCtx context.Context, i [][]byte)
 	m.logger.Debug("getting iterators")
 
 	pendingFilter := txpool.PendingFilter{
-		MinTip:       nil,
+		MinTip:       m.minTip,
 		BaseFee:      baseFeeUint,
 		BlobFee:      nil,
 		OnlyPlainTxs: true,